Rename data/index.ts into GovernCore
Also move it into src/core as it is now the main interface of the core
library.
