+++ packages/govern-server/src/core/index.ts
@@ -1,0 +1,1 @@
+export { default as GovernCore } from './GovernCore'
