+++ packages/govern-server/src/index.ts
@@ -1,9 +1,9 @@
 import env from './env'
 import startGraphql from './api-graphql'
-import GovernData from './core/data'
+import { GovernCore } from './core'
 
 async function main() {
-  const govern = new GovernData({ network: 4 })
+  const govern = new GovernCore({ network: 4 })
 
   const { url: graphqlUrl } = await startGraphql({
     govern,
